textmock => text2[type]
